Safari-611.1.10/LayoutTests/http/tests/security/contentSecurityPolicy/image-blocked-expected.txt

CONSOLE MESSAGE: Refused to load http://127.0.0.1:8000/security/resources/abe.png because it does not appear in the img-src directive of the Content Security Policy.
ALERT: PASS
This test passes if it doesn't alert FAIL and does alert PASS.
